devops script working with local

+++ devops.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-#show errors with exit code
 set -euo pipefail
 COMPOSE="docker-compose"
 COPIED=false
@@ -7,21 +6,31 @@ export LOC=${LOC:-local}
 ABORT=true
 
 function install() {
-  if [ "$LOC" == "ci" ];then
-    if [ ! -f docker-compose.yml ] && [ ! -f vote_app_backend/vote_app_backend/settings.py ];then
-      docker network create proxy
-      cp .deploy/${LOC}/docker-compose.yml docker-compose.yml || true
-      cp .deploy/${LOC}/settings.py vote_app_backend/vote_app_backend/settings.py || true
-      COPIED=true
-    fi
-  fi
-  if [ "$LOC" == "local" ] || [ "$LOC" == "stage" ];then
-    #if file does not exist
+  if [ "$LOC" == "local" ];then
+    docker network create proxy
     if [ ! -f docker-compose.yml ] && [ ! -f frontend-ui/vue.config.js ] && [ ! -f vote_app_backend/vote_app_backend/settings.py ];then
       cp .deploy/${LOC}/docker-compose.yml docker-compose.yml || true
       cp .deploy/${LOC}/vue.config.js frontend-ui/vue.config.js || true
       cp .deploy/${LOC}/settings.py vote_app_backend/vote_app_backend/settings.py || true
       COPIED=true
+      ABORT=false
+    fi
+  fi
+  if [ "$LOC" == "ci" ];then
+    if [ ! -f docker-compose.yml ] && [ ! -f vote_app_backend/vote_app_backend/settings.py ];then
+      cp .deploy/${LOC}/docker-compose.yml docker-compose.yml || true
+      cp .deploy/${LOC}/settings.py vote_app_backend/vote_app_backend/settings.py || true
+      COPIED=true
+      ABORT=false
+    fi
+  fi
+  if [ "$LOC" == "stage" ];then
+    if [ ! -f docker-compose.yml ] && [ ! -f frontend-ui/vue.config.js ] && [ ! -f vote_app_backend/vote_app_backend/settings.py ];then
+      cp .deploy/${LOC}/docker-compose.yml docker-compose.yml || true
+      cp .deploy/${LOC}/vue.config.js frontend-ui/vue.config.js || true
+      cp .deploy/${LOC}/settings.py vote_app_backend/vote_app_backend/settings.py || true
+      COPIED=true
+      ABORT=false
     fi
   fi
 }
